+++ src/main/php/response/ResponseNegotiator.php
@@ -54,12 +54,12 @@ class ResponseNegotiator
      * response in case a mime type but no suitable formatter was found.
      *
      * @param   \stubbles\input\web\WebRequest                $request
-     * @param   \stubbles\webapp\response\SupportedMimeTypes  $supportedMimeTypes
+     * @param   \stubbles\webapp\response\SupportedMimeTypes  $supportedMimeTypes  optional
      * @return  Response
      */
-    public function negotiateMimeType(WebRequest $request, SupportedMimeTypes $supportedMimeTypes)
+    public function negotiateMimeType(WebRequest $request, SupportedMimeTypes $supportedMimeTypes = null)
     {
-        if ($supportedMimeTypes->isContentNegotationDisabled()) {
+        if (null === $supportedMimeTypes || $supportedMimeTypes->isContentNegotationDisabled()) {
             return $this->response;
         }